requirements.txt
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.2
